# apispec/EventApp-APIspec.yaml
openapi: 3.0.3
info:
  title: Events App API
  description: |
    Events Management Platform API - MVP Faza 1

    Omogućava korisnicima kreiranje i upravljanje događajima sa QR kod sistemom
    za upload slika, upravljanje gostima i kolaboraciju.

    **Autentifikacija:** AWS Cognito + Gmail OAuth
    **Regioni:** Srbija, BiH, Hrvatska, Crna Gora, Makedonija
  version: 1.0.0
  contact:
    name: Events App API Support
    email: support@eventsapp.com
  license:
    name: MIT
  x-gitbook-description-document:
    object: document
    data:
      schemaVersion: 8
    nodes:
      - object: block
        type: paragraph
        isVoid: false
        data: {}
        nodes:
          - object: text
            leaves:
              - object: leaf
                text: Events Management Platform API - MVP Faza 1
                marks: []
      - object: block
        type: paragraph
        isVoid: false
        data: {}
        nodes:
          - object: text
            leaves:
              - object: leaf
                text: >-
                  Omogućava korisnicima kreiranje i upravljanje događajima sa QR
                  kod sistemom za upload slika, upravljanje gostima i
                  kolaboraciju.
                marks: []
      - object: block
        type: paragraph
        isVoid: false
        data: {}
        nodes:
          - object: text
            leaves:
              - object: leaf
                text: 'Autentifikacija:'
                marks:
...
servers:
...
security:
  - CognitoAuth: []
...
paths:
...
  /users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          x-gitbook-description-html: <p>User profile retrieved</p>
    put:
      tags:
        - Users
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          x-gitbook-description-html: <p>Profile updated successfully</p>
  /events:
    get:
      tags:
        - Events
      summary: Get user's events
      description: Retrieve all events created by or shared with the user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              - active
              - expired
              - draft
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  total:
                    type: integer
                  has_more:
                    type: boolean
          x-gitbook-description-html: <p>Events retrieved successfully</p>
      x-gitbook-description-document:
        object: document
        data:
          schemaVersion: 8
        nodes:
          - object: block
            type: paragraph
            isVoid: false
            data: {}
            nodes:
              - object: text
                leaves:
                  - object: leaf
                    text: Retrieve all events created by or shared with the user
                    marks: []
      x-gitbook-description-html: <p>Retrieve all events created by or shared with the user</p>
    post:
      tags:
        - Events
      summary: Create new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          x-gitbook-description-html: <p>Event created successfully</p>
        '400':
          description: Invalid input data
          x-gitbook-description-html: <p>Invalid input data</p>
        '402':
          description: Payment required for selected plan
          x-gitbook-description-html: <p>Payment required for selected plan</p>
  /events/{event_id}:
    get:
      tags:
        - Events
      summary: Get event details
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDetail'
          x-gitbook-description-html: <p>Event details retrieved</p>
        '404':
          description: Event not found
          x-gitbook-description-html: <p>Event not found</p>
    put:
      tags:
        - Events
      summary: Update event
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdate'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          x-gitbook-description-html: <p>Event updated successfully</p>
    delete:
      tags:
        - Events
      summary: Delete event
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Event deleted successfully
          x-gitbook-description-html: <p>Event deleted successfully</p>
        '403':
          description: Not authorized to delete this event
          x-gitbook-description-html: <p>Not authorized to delete this event</p>
...
  /events/{event_id}/agenda:
    get:
      tags:
        - Agenda
      summary: Get event agenda
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Agenda retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agenda'
          x-gitbook-description-html: <p>Agenda retrieved successfully</p>
        '404':
          description: Agenda not found
          x-gitbook-description-html: <p>Agenda not found</p>
    post:
      tags:
        - Agenda
      summary: Create event agenda
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgendaCreate'
      responses:
        '201':
          description: Agenda created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agenda'
          x-gitbook-description-html: <p>Agenda created successfully</p>
    put:
      tags:
        - Agenda
      summary: Update event agenda
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgendaUpdate'
      responses:
        '200':
          description: Agenda updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agenda'
          x-gitbook-description-html: <p>Agenda updated successfully</p>
    delete:
      tags:
        - Agenda
      summary: Delete event agenda
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Agenda deleted successfully
          x-gitbook-description-html: <p>Agenda deleted successfully</p>
  /events/{event_id}/agenda/items:
    post:
      tags:
        - Agenda
      summary: Add agenda item
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgendaItemCreate'
      responses:
        '201':
          description: Agenda item added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgendaItem'
          x-gitbook-description-html: <p>Agenda item added successfully</p>
  /events/{event_id}/agenda/items/{item_id}:
    put:
      tags:
        - Agenda
      summary: Update agenda item
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgendaItemUpdate'
      responses:
        '200':
          description: Agenda item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgendaItem'
          x-gitbook-description-html: <p>Agenda item updated successfully</p>
    delete:
      tags:
        - Agenda
      summary: Delete agenda item
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Agenda item deleted successfully
          x-gitbook-description-html: <p>Agenda item deleted successfully</p>
  /events/{event_id}/agenda/reorder:
    put:
      tags:
        - Agenda
      summary: Reorder agenda items
      description: Bulk update display_order values for multiple agenda items with atomic transaction handling
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgendaReorderRequest'
      responses:
        '200':
          description: Agenda items reordered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Agenda items successfully reordered."
          x-gitbook-description-html: <p>Agenda items reordered successfully</p>
        '400':
          description: Some agenda items don't belong to the specified agenda
          x-gitbook-description-html: <p>Some agenda items don't belong to the specified agenda</p>
        '403':
          description: User doesn't have permission to access this event
          x-gitbook-description-html: <p>User doesn't have permission to access this event</p>
        '404':
          description: Event or agenda not found
          x-gitbook-description-html: <p>Event or agenda not found</p>
...
components:
  securitySchemes:
    CognitoAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT token
      x-gitbook-description-html: <p>AWS Cognito JWT token</p>
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
    EventPlan:
      type: string
      enum:
        - freemium
        - starter
        - plus
        - full
      description: |
        - freemium: 50 slika, 15 dana, ZIP download
        - starter: 200 slika, 3 meseca, ZIP download
        - plus: 500 slika, 6 meseci, galerija, 2 kolaboratora
        - full: 2000 slika, 12 meseci, galerija, 5 kolaboratora
      x-gitbook-description-html: |-
        <ul>
        <li>freemium: 50 slika, 15 dana, ZIP download</li>
        <li>starter: 200 slika, 3 meseca, ZIP download</li>
        <li>plus: 500 slika, 6 meseci, galerija, 2 kolaboratora</li>
        <li>full: 2000 slika, 12 meseci, galerija, 5 kolaboratora</li>
        </ul>
    EventCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        plan:
          $ref: '#/components/schemas/EventPlan'
        location:
          type: string
        restaurant_name:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
          format: time
        event_type:
          type: string
          enum:
            - wedding
            - birthday
            - baptism
            - graduation
            - anniversary
            - corporate
            - other
        expected_guests:
          type: integer
          minimum: 1
        description:
          type: string
          maxLength: 1000
      required:
        - name
        - plan
        - location
        - date
        - time
        - event_type
    EventUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        location:
          type: string
        restaurant_name:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
          format: time
        event_type:
          type: string
          enum:
            - wedding
            - birthday
            - baptism
            - graduation
            - anniversary
            - corporate
            - other
        expected_guests:
          type: integer
          minimum: 1
        description:
          type: string
          maxLength: 1000
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        plan:
          $ref: '#/components/schemas/EventPlan'
        location:
          type: string
        restaurant_name:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
          format: time
        event_type:
          type: string
        expected_guests:
          type: integer
        description:
          type: string
        qr_code_url:
          type: string
          format: uri
        landing_page_url:
          type: string
          format: uri
        photo_count:
          type: integer
        guest_count:
          type: integer
        status:
          type: string
          enum:
            - active
            - expired
            - draft
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        owner:
          $ref: '#/components/schemas/User'
    EventDetail:
      allOf:
        - $ref: '#/components/schemas/Event'
        - type: object
          properties:
            collaborators:
              type: array
              items:
                $ref: '#/components/schemas/Collaborator'
            landing_config:
              $ref: '#/components/schemas/LandingPageConfig'
            menu:
              $ref: '#/components/schemas/Menu'
            agenda:
              $ref: '#/components/schemas/Agenda'
            plan_limits:
              type: object
              properties:
                max_photos:
                  type: integer
                max_collaborators:
                  type: integer
                access_days:
                  type: integer
                has_gallery:
                  type: boolean
    LandingPageConfig:
      type: object
      properties:
        require_uploader_info:
          type: boolean
          default: false
        sections_order:
          type: array
          items:
            type: string
            enum:
              - upload
              - menu
              - agenda
          default:
            - upload
            - menu
            - agenda
        custom_message:
          type: string
          maxLength: 500
        show_event_details:
          type: boolean
          default: true
...
    Collaborator:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        permissions:
          type: array
          items:
            type: string
            enum:
              - manage_guests
              - manage_photos
              - edit_event
              - manage_landing
              - manage_menu
              - manage_agenda
        added_at:
          type: string
          format: date-time
        added_by:
          $ref: '#/components/schemas/User'
...
    Menu:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/MenuCategory'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    MenuCategory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        display_order:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    MenuItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: decimal
        currency:
          type: string
        allergens:
          type: array
          items:
            type: string
        dietary_info:
          type: array
          items:
            type: string
        display_order:
          type: integer
        available:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgendaCreate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          default: Program događaja
        description:
          type: string
          maxLength: 500
        items:
          type: array
          items:
            $ref: '#/components/schemas/AgendaItemCreate'
    AgendaUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 500
    Agenda:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/AgendaItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgendaItemCreate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 500
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        location:
          type: string
          maxLength: 200
        type:
          type: string
          enum:
            - ceremony
            - reception
            - entertainment
            - speech
            - meal
            - break
            - photo_session
            - other
        display_order:
          type: integer
          default: 0
        is_important:
          type: boolean
          default: false
          description: Highlight important events (ceremony, speeches, etc.)
          x-gitbook-description-html: <p>Highlight important events (ceremony, speeches, etc.)</p>
      required:
        - title
        - start_time
    AgendaItemUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 500
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        location:
          type: string
          maxLength: 200
        type:
          type: string
          enum:
            - ceremony
            - reception
            - entertainment
            - speech
            - meal
            - break
            - photo_session
            - other
        display_order:
          type: integer
        is_important:
          type: boolean
    AgendaItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        location:
          type: string
        type:
          type: string
        display_order:
          type: integer
        is_important:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgendaReorderRequest:
      type: object
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ReorderItem'
          description: List of items with new display orders
      required:
        - items
    ReorderItem:
      type: object
      properties:
        item_id:
          type: string
          format: uuid
          description: ID of the agenda item
        display_order:
          type: integer
          minimum: 0
          description: New display order for the item
      required:
        - item_id
        - display_order